[MODEL-8452] Move unit tests to their own file
